refactor: extract BudgetBoard component and add login validation
